refactor: modularize the `standalone` module
